+++ src/main/java/com/ossanasur/cbconnect/module/courrier/entity/Courrier.java
@@ -61,7 +61,7 @@ public class Courrier extends InternalHistorique {
     private LocalDateTime dateTraitement;
 
     @Column(name = "ossan_ged_document_id")
-    private Long ossanGedDocumentId;
+    private Integer ossanGedDocumentId;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "sinistre_id")
